[#243] Added constructor and create methods to TeamMembersFormBase.
