+++ modules/apigee_edge_teams/src/Form/TeamMembersFormBase.php
@@ -21,7 +21,9 @@
 namespace Drupal\apigee_edge_teams\Form;
 
 use Drupal\apigee_edge_teams\Entity\TeamRoleInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormBase;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Base team members form.
@@ -48,6 +50,26 @@ abstract class TeamMembersFormBase extends FormBase {
    * @var \Drupal\apigee_edge_teams\Entity\Storage\TeamMemberRoleStorage
    */
   protected $teamMemberRoleStorage;
+
+  /**
+   * TeamMembersFormBase constructor.
+   *
+   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
+   *   The entity type manager.
+   */
+  public function __construct(EntityTypeManagerInterface $entity_type_manager) {
+    $this->teamRoleStorage = $entity_type_manager->getStorage('team_role');
+    $this->teamMemberRoleStorage = $entity_type_manager->getStorage('team_member_role');
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container) {
+    return new static(
+      $container->get('entity_type.manager')
+    );
+  }
 
   /**
    * Returns an array of team role options keyed by team role id.
